feat: add ViewWithMethodsClassReflectionExtension
